Refactor application to use DTOs, exception handling and API testing.

+++ src/main/java/com/loan/dao/LoanDAO.java
@@ -41,7 +41,7 @@ public class LoanDAO {
 		em.close();
 	}
 
-	public Loan findById(int loanId) {
+	public Loan findById(Long loanId) {
 
 		EntityManager em = emf.createEntityManager();
